refactor(upgrades): keep package execution in adapter boundary
